sso-proxy: refactor reverse proxy into separate file w/ tests
